Created version of easykey with no jquery requirement
The function signature changed from

    $.onKey(whichKey, onKey, options)

to

    easyKey.onKey(whichKey, nodeCollection, onKey, options)

where nodeCollection can be a single Node, a NodeList or an Array of Nodes.

Unfortunately, the calls aren't chainable anymore, which was a big benefit
of the jQuery version. I'm sure this could be improved to be chainable, but
for a quick-fix, this removes the jQuery requirement.